Estimation of the Area Under Prescription Evaluation Curve (AUPEC) in Randomized Experiments

This function estimates AUPEC. The details of the methods for this design are given in Imai and Li (2019).

Usage

AUPEC(T, tau, Y, centered = TRUE)

Arguments

T

A vector of the unit-level binary treatment receipt variable for each sample.

tau

A vector of the unit-level continuous score for treatment assignment. We assume those that have tau<0 should not have treatment. Conditional Average Treatment Effect is one possible measure.

Y

A vector of the outcome variable of interest for each sample.

centered

If TRUE, the outcome variables would be centered before processing. This minimizes the variance of the estimator. Default is TRUE.

Value

A list that contains the following items:

aupec

The estimated Area Under Prescription Evaluation Curve

sd

The estimated standard deviation of AUPEC.

vec

A vector of points outlining the AUPEC curve across each possible budget point for the dataset. Each step increases the budget by 1/n where n is the number of data points.

References

Imai and Li (2019). “Experimental Evaluation of Individualized Treatment Rules”,

Author

Michael Lingzhi Li, Operations Research Center, Massachusetts Institute of Technology mlli@mit.edu, http://mlli.mit.edu;

Examples

T = c(1,0,1,0,1,0,1,0)
tau = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7)
Y = c(4,5,0,2,4,1,-4,3)
aupeclist <- AUPEC(T,tau,Y)
aupeclist$aupec
#> [1] -0.5820312
aupeclist$sd
#> [1] 0.7675369
aupeclist$vec
#> [1]  0.59375 -0.87500 -0.65625 -0.12500 -0.15625 -0.62500 -1.40625 -1.40625